repos/smart-contract-tx-rollup/src/rollup.py

class TxRollup:
    def __init__(self, w3, contract):
        self.w3 = w3
        self.contract = contract
        self.approved_transactions = []
        self.rollup_batch = []

    def fetch_approved_transactions(self):
        # Fetch approved transactions from the smart contract
        # This is a placeholder and should be implemented based on your contract's structure
        self.approved_transactions = self.contract.functions.getApprovedTransactions().call()

    def aggregate_transactions(self):
        # Aggregate transactions into a batch
        # This is a simple aggregation, you might want to optimize based on gas costs, etc.
        self.rollup_batch = self.approved_transactions

    def submit_rollup(self):
        # Submit the rollup batch to the blockchain
        # This is a placeholder and should be implemented based on your contract's structure
        tx_hash = self.contract.functions.submitRollup(self.rollup_batch).transact()
        self.w3.eth.wait_for_transaction_receipt(tx_hash)
        print(f"Rollup submitted. Transaction hash: {tx_hash.hex()}")
