repos/smart-contract-tx-rollup/tests/test_rollup.py

import pytest
from web3 import Web3
from src.rollup import TxRollup

@pytest.fixture
def w3():
    return Web3(Web3.EthereumTesterProvider())

@pytest.fixture
def contract(w3):
    # Deploy a mock contract for testing
    # This is a placeholder and should be implemented based on your contract's structure
    return w3.eth.contract(abi=[], bytecode="0x")

def test_fetch_approved_transactions(w3, contract):
    rollup = TxRollup(w3, contract)
    rollup.fetch_approved_transactions()
    assert isinstance(rollup.approved_transactions, list)

def test_aggregate_transactions(w3, contract):
    rollup = TxRollup(w3, contract)
    rollup.approved_transactions = [1, 2, 3]
    rollup.aggregate_transactions()
    assert rollup.rollup_batch == [1, 2, 3]

def test_submit_rollup(w3, contract):
    rollup = TxRollup(w3, contract)
    rollup.rollup_batch = [1, 2, 3]
    # This test is a placeholder and should be implemented based on your contract's structure
    # rollup.submit_rollup()
    # assert some condition about the submitted rollup
